requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==45.0.7
daphne==4.2.1
dj-rest-auth==7.0.1
Django==5.2.6
django-allauth==65.11.2
django-celery-beat==2.8.1
django-cors-headers==4.8.0
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.16.1
Faker==37.6.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.4
Markdown==3.9
numpy==2.3.3
packaging==25.0
pandas==2.3.2
prompt_toolkit==3.0.52
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.2.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.4.0
requests==2.32.5
service-identity==24.2.0
setuptools==78.1.1
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
wheel==0.45.1
zope.interface==8.0
